Renamed estimate usage
