+++ pkg/estimate/class.go
@@ -1,4 +1,4 @@
-package reports
+package estimate
 
 type Class struct {
