+++ pkg/estimate/report.go
@@ -1,4 +1,4 @@
-package reports
+package estimate
 
 type Report struct {
