+++ pkg/networks/network.go
@@ -2,11 +2,11 @@ package networks
 
 import (
 	"gonum.org/v1/gonum/mat"
-	"iio/pkg/reports"
+	"iio/pkg/estimate"
 )
 
 type Network interface {
 	Train()
-	Test() *reports.Report
+	Test() *estimate.Report
 	Evaluate(mat.Vector) int
 }
